app/requirements.txt
anvil-app-server==1.10.1
beautifulsoup4==4.12.2
bs4==0.0.1
bz2file==0.98
ciseau==1.0.1
epub==0.5.2
et-xmlfile==1.1.0
openpyxl==3.1.2
pyodbc==5.0.1
soupsieve==2.5
xml_cleaner==2.0.4
pyodbc-5.0.1
openpyxl-3.1.2
epub_conversion-1.0.15